Updated uint64_t/int64_t to usize/isize

+++ src/main.c
@@ -11,7 +11,7 @@ int main(int argc, const char* argv[]) {
         path = argv[1];
     }
 
-    int64_t dd = open_directory(path);
+    isize dd = open_directory(path);
     if (dd < 0) {
         printf("Error while opening %s: %s\n", argv[1], strerror(dd));
         exit(1);
@@ -19,7 +19,7 @@ int main(int argc, const char* argv[]) {
 
     Dirent entry;
     while (1) {
-        int64_t status = read_directory(dd, &entry);
+        isize status = read_directory(dd, &entry);
         if (status == 0) {
             break;
         } else if (status < 0) {
